HG-330 Always show caption in lightbox
This change affects lightbox media in landscape view. Currently the
caption is hidden by an image if that image is resized large enough to
cover it. Here, the z-index of the caption is increased to match the
header (which keeps the close X viewable). Consequently, the caption
area will cover a portion of the bottom of the image in this situation.
This commit also fixes a JS error when resizing images in lightbox view.