HG-339 Fix transition freezes on iOS8
https://wikia-inc.atlassian.net/browse/HG-339

These changes attempt to fix article transition freezes on iOS8 Safari.
HammerJS is not needed for link clicking, nor is Fastclick.

If we continue to use Fastclick, we will most likely use our own fork
due to the change needed to fix our particular issue with it. The
version in bower config is at this time _temporary_ while being tested.