HG-515 Fix trackClick action for components
This “bug” was caused by a recent EmberJS upgrade. However, the “bug”
is actually not a bug but a hack on our side. See this link for more
information:
https://github.com/emberjs/ember.js/issues/9541

Actions referenced in components need to be handled by the component,
and will no longer “bubble up” to be handled by the parent
component/controller/route.