Enable suppressImplicitAnyIndexErrors flag
This change would enable the ```—suppressImplicitAnyIndexErrors``` flag
for ```tsc```. This Github issue provides some background about how the
flag came about: https://github.com/Microsoft/TypeScript/issues/1232

I’m encouraging this change because I think in many cases, the types
for string indices are obvious to us as developers and are not
necessary. And in some cases, providing typing for these can require
rearranging and adding code, making it tedious at least and at worst a
waste of time. In other words, a benefit that Typescript provides is
actually working against us.

Here’s a real-world example where this would apply.
```
private getCredentials (): LoginCredentials {
    return {
        username: this.form.elements['username'].value,
        password: this.form.elements['password'].value
    };
}
```
This is a member function of a Login class. ```this.form``` has a type
of ```HTMLFormElement``` which has the property ```elements```, which
in turn can be accessed by an index that is the id or name of a form
element. Without the ```—suppressImplicitAnyIndexErrors``` flag, you
would see compilation warnings like:
```7017 Index signature of object type implicitly has an 'any' type.```
and you’d need to create an interface or do some casting to satisfy the
compiler. With the flag enabled, these kinds of cases are ignored.