INT-102: Add photo to section in Japanese Wikia.
Show photo icon to each section in Japanese Wikia. When a user clicks the icon, then the user can choose an image file from photo gallery and upload the file. When upload is successful, then we add the uploaded file reference to wiki text at the bottom of the section.