SOC-1122 Always show the spinner if there are more items to load.
This fixes some problems on iOS Chrome where JS won't run while the content is being scrolled.  Having his always be in the content means that the spinner will show up when the user scrolls to the bottom, even if it doesn't start spinning until the user lifts their finger.