(DAT-4579) Treat local URLs with query parameters as external URLs
If a local article URL has a query param, we should treat it as a full external
URL if it is not handled elsewhere. Manual local URLs with query params created by
users are already handled as the MediaWiki parser adds the external class which
is already checked in Mercury.

This is an extra safety check in case any extensions add URLs with query params
(as TabView did) so that they don't result in a broken page when clicked.